refactor: refactor for code convention
