+++ src/main/java/webserver/RequestHandler.java
@@ -9,7 +9,7 @@ import http.request.HttpRequest;
 import http.request.HttpRequestReader;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import utils.ControllerGenerator;
+import controller.ControllerGenerator;
 
 public class RequestHandler implements Runnable {
 	private static final Logger logger = LoggerFactory.getLogger(RequestHandler.class);
